client/requirements.txt
python-socketio[client]
websocket-client
mss
Pillow
dxcam
numpy
pyinstaller
